+++ packages/edition-slices/src/tiles/tile-u/styles/index.js
@@ -4,30 +4,83 @@ import {
   editionBreakpoints
 } from "@times-components/styleguide";
 
-const styles = fontSize => ({
+const sharedStyles = {
   container: {
     flex: 1,
-    flexDirection: "row",
-    padding: spacing(2)
+    flexDirection: "row"
   },
   headline: {
     ...fontFactory({
       font: "headline",
-      fontSize
-    }),
-    paddingVertical: spacing(1)
+      fontSize: "tileLeadHeadline"
+    })
   },
   imageContainer: {
     width: "60%"
   },
   summaryContainer: {
-    paddingBottom: spacing(1),
-    paddingRight: spacing(4),
     width: "40%"
   }
-});
+};
 
-export default breakpoint =>
-  editionBreakpoints.huge === breakpoint
-    ? styles("articleHeadline")
-    : styles("tileLeadHeadline");
+const mediumBreakpointStyles = {
+  ...sharedStyles,
+  container: {
+    ...sharedStyles.container,
+    paddingVertical: spacing(3),
+    paddingRight: spacing(2)
+  },
+  headline: {
+    ...sharedStyles.headline,
+    fontSize: 30,
+    lineHeight: 30,
+    marginBottom: spacing(2)
+  },
+  summaryContainer: {
+    ...sharedStyles.summaryContainer,
+    paddingRight: spacing(4)
+  },
+  star: {
+    starButton: {
+      position: "absolute",
+      right: spacing(3),
+      bottom: 0
+    }
+  }
+};
+
+const wideBreakpointStyles = {
+  ...sharedStyles,
+  container: {
+    ...sharedStyles.container,
+    padding: spacing(2)
+  },
+  headline: {
+    ...sharedStyles.headline,
+    paddingVertical: spacing(1)
+  },
+  summaryContainer: {
+    ...sharedStyles.summaryContainer,
+    paddingBottom: spacing(1),
+    paddingRight: spacing(4)
+  }
+};
+
+const hugeBreakpointStyles = {
+  ...wideBreakpointStyles,
+  headline: {
+    ...wideBreakpointStyles.headline,
+    ...fontFactory({
+      font: "headline",
+      fontSize: "articleHeadline"
+    })
+  }
+};
+
+const stylesResolver = {
+  [editionBreakpoints.medium]: mediumBreakpointStyles,
+  [editionBreakpoints.wide]: wideBreakpointStyles,
+  [editionBreakpoints.huge]: hugeBreakpointStyles
+};
+
+export default breakpoint => stylesResolver[breakpoint];