+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.99](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.1.98...@times-components/article-byline@3.1.99) (2019-07-24)
+
+**Note:** Version bump only for package @times-components/article-byline
+
+
+
+
+
 ## [3.1.98](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.1.97...@times-components/article-byline@3.1.98) (2019-07-22)
 
 **Note:** Version bump only for package @times-components/article-byline