+++ packages/save-star-web/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.3.2...@times-components/save-star-web@0.3.3) (2019-07-24)
+
+
+### Bug Fixes
+
+* url in share mail ([#2129](https://github.com/newsuk/times-components/issues/2129)) ([9fbc1fe](https://github.com/newsuk/times-components/commit/9fbc1fe))
+
+
+
+
+
 ## [0.3.2](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.3.1...@times-components/save-star-web@0.3.2) (2019-07-19)
 
 **Note:** Version bump only for package @times-components/save-star-web