+++ storybook-native/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.18](https://github.com/newsuk/times-components/compare/@times-components/storybook-native@0.0.17...@times-components/storybook-native@0.0.18) (2019-07-24)
+
+
+### Bug Fixes
+
+* url in share mail ([#2129](https://github.com/newsuk/times-components/issues/2129)) ([9fbc1fe](https://github.com/newsuk/times-components/commit/9fbc1fe))
+
+
+
+
+
 ## [0.0.17](https://github.com/newsuk/times-components/compare/@times-components/storybook-native@0.0.16...@times-components/storybook-native@0.0.17) (2019-07-19)
 
 **Note:** Version bump only for package @times-components/storybook-native